chore: decouple methods into individual fs modules (#50)
