+++ src/methods/adversarial/mod.rs
@@ -1,0 +1,22 @@
+use super::*;
+
+pub mod check_store;
+pub use check_store as adv_check_store;
+
+pub mod disable_doomslug;
+pub use disable_doomslug as adv_disable_doomslug;
+
+pub mod disable_header_sync;
+pub use disable_header_sync as adv_disable_header_sync;
+
+pub mod get_saved_blocks;
+pub use get_saved_blocks as adv_get_saved_blocks;
+
+pub mod produce_blocks;
+pub use produce_blocks as adv_produce_blocks;
+
+pub mod set_weight;
+pub use set_weight as adv_set_weight;
+
+pub mod switch_to_height;
+pub use switch_to_height as adv_switch_to_height;
