+++ src/methods/mod.rs
@@ -1,0 +1,220 @@
+use std::io;
+
+use serde::Deserialize;
+use serde_json::json;
+use thiserror::Error;
+
+mod private {
+    pub trait Sealed {}
+}
+
+pub trait RpcMethod: private::Sealed
+where
+    Self::Response: RpcHandlerResponse,
+    Self::Error: RpcHandlerError,
+{
+    type Response;
+    type Error;
+
+    fn method_name(&self) -> &str;
+
+    fn params(&self) -> Result<serde_json::Value, io::Error>;
+}
+
+impl<T> private::Sealed for &T where T: private::Sealed {}
+impl<T> RpcMethod for &T
+where
+    T: RpcMethod,
+{
+    type Response = T::Response;
+    type Error = T::Error;
+
+    fn method_name(&self) -> &str {
+        T::method_name(self)
+    }
+
+    fn params(&self) -> Result<serde_json::Value, io::Error> {
+        T::params(self)
+    }
+}
+
+pub trait RpcHandlerResponse: serde::de::DeserializeOwned {
+    fn parse(value: serde_json::Value) -> Result<Self, serde_json::Error> {
+        serde_json::from_value(value)
+    }
+}
+
+pub trait RpcHandlerError: serde::de::DeserializeOwned {
+    /// Parser for the `.error_struct` field in RpcError.
+    fn parse(handler_error: serde_json::Value) -> Result<Self, serde_json::Error> {
+        serde_json::from_value(handler_error)
+    }
+
+    /// Parser for the `.data` field in RpcError, not `.error_struct`.
+    ///
+    /// This would only ever be used as a fallback if [`RpcHandlerError::parse`] fails.
+    ///
+    /// Defaults to `None` meaning there's no alternative deserialization available.
+    fn parse_raw_error(_error: serde_json::Value) -> Option<Result<Self, serde_json::Error>> {
+        None
+    }
+}
+
+pub mod block;
+pub mod broadcast_tx_async;
+pub mod broadcast_tx_commit;
+pub mod chunk;
+pub mod gas_price;
+pub mod health;
+pub mod light_client_proof;
+pub mod network_info;
+pub mod next_light_client_block;
+pub mod query;
+pub mod status;
+pub mod tx;
+pub mod validators;
+
+// ======== experimental ========
+mod experimental;
+pub use experimental::EXPERIMENTAL_changes;
+pub use experimental::EXPERIMENTAL_changes_in_block;
+pub use experimental::EXPERIMENTAL_check_tx;
+pub use experimental::EXPERIMENTAL_genesis_config;
+pub use experimental::EXPERIMENTAL_protocol_config;
+pub use experimental::EXPERIMENTAL_receipt;
+pub use experimental::EXPERIMENTAL_tx_status;
+pub use experimental::EXPERIMENTAL_validators_ordered;
+// ======== experimental ========
+
+// ======== any ========
+#[cfg(feature = "any")]
+mod any;
+#[cfg(feature = "any")]
+pub use any::{request as any, RpcAnyRequest};
+// ======== any ========
+
+// ======== sandbox ========
+#[cfg(feature = "sandbox")]
+mod sandbox;
+
+#[cfg(feature = "sandbox")]
+pub use sandbox::sandbox_patch_state;
+
+#[cfg(feature = "sandbox")]
+pub use sandbox::sandbox_fast_forward;
+// ======== sandbox ========
+
+// ======== adversarial ========
+#[cfg(feature = "adversarial")]
+mod adversarial;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_set_weight;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_disable_header_sync;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_disable_doomslug;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_produce_blocks;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_switch_to_height;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_get_saved_blocks;
+
+#[cfg(feature = "adversarial")]
+pub use adversarial::adv_check_store;
+// ======== adversarial ========
+
+mod common {
+    use super::*;
+
+    // workaround for deserializing partially serialized
+    // error types missing the `error_message` field in
+    // their UnknownBlock variants.
+    macro_rules! _parse_unknown_block {
+        ($json:expr => $err_ty:ident) => {
+            if $json["name"] == "UNKNOWN_BLOCK" {
+                Some(Ok($err_ty::UnknownBlock {
+                    error_message: "".to_string(),
+                }))
+            } else {
+                None
+            }
+        };
+    }
+    pub(crate) use _parse_unknown_block as parse_unknown_block;
+
+    pub fn serialize_signed_transaction(
+        tx: &near_primitives::transaction::SignedTransaction,
+    ) -> Result<String, io::Error> {
+        Ok(near_primitives::serialize::to_base64(
+            &borsh::BorshSerialize::try_to_vec(&tx)?,
+        ))
+    }
+
+    // adv_*
+    #[cfg(feature = "adversarial")]
+    impl RpcHandlerError for () {}
+
+    // adv_*
+    #[cfg(feature = "adversarial")]
+    impl RpcHandlerResponse for () {
+        fn parse(_value: serde_json::Value) -> Result<Self, serde_json::Error> {
+            Ok(())
+        }
+    }
+
+    #[cfg(feature = "any")]
+    impl RpcHandlerResponse for serde_json::Value {
+        fn parse(value: serde_json::Value) -> Result<Self, serde_json::Error> {
+            Ok(value)
+        }
+    }
+
+    #[cfg(feature = "any")]
+    impl RpcHandlerError for serde_json::Value {
+        fn parse(handler_error: serde_json::Value) -> Result<Self, serde_json::Error> {
+            Ok(handler_error)
+        }
+    }
+
+    // broadcast_tx_commit, tx
+    impl RpcHandlerResponse for near_primitives::views::FinalExecutionOutcomeView {}
+
+    // broadcast_tx_commit, tx, EXPERIMENTAL_check_tx, EXPERIMENTAL_tx_status
+    impl RpcHandlerError for near_jsonrpc_primitives::types::transactions::RpcTransactionError {
+        fn parse_raw_error(value: serde_json::Value) -> Option<Result<Self, serde_json::Error>> {
+            match serde_json::from_value::<near_jsonrpc_primitives::errors::ServerError>(value) {
+                Ok(near_jsonrpc_primitives::errors::ServerError::TxExecutionError(
+                    near_primitives::errors::TxExecutionError::InvalidTxError(context),
+                )) => Some(Ok(Self::InvalidTransaction { context })),
+                Err(err) => Some(Err(err)),
+                _ => None,
+            }
+        }
+    }
+
+    // health, status
+    impl RpcHandlerError for near_jsonrpc_primitives::types::status::RpcStatusError {}
+
+    // EXPERIMENTAL_changes, EXPERIMENTAL_changes_in_block
+    impl RpcHandlerError for near_jsonrpc_primitives::types::changes::RpcStateChangesError {
+        fn parse_raw_error(value: serde_json::Value) -> Option<Result<Self, serde_json::Error>> {
+            parse_unknown_block!(value => Self)
+        }
+    }
+
+    // EXPERIMENTAL_broadcast_tx_sync, EXPERIMENTAL_check_tx
+    impl RpcHandlerResponse
+        for near_jsonrpc_primitives::types::transactions::RpcBroadcastTxSyncResponse
+    {
+    }
+
+    // validators, EXPERIMENTAL_validators_ordered
+    impl RpcHandlerError for near_jsonrpc_primitives::types::validator::RpcValidatorError {}
+}
