+++ src/methods/next_light_client_block.rs
@@ -1,0 +1,31 @@
+use super::*;
+
+pub use near_jsonrpc_primitives::types::light_client::{
+    RpcLightClientNextBlockError, RpcLightClientNextBlockRequest,
+};
+pub use near_primitives::views::LightClientBlockView;
+
+pub type RpcLightClientNextBlockResponse = Option<LightClientBlockView>;
+
+impl RpcHandlerResponse for RpcLightClientNextBlockResponse {}
+
+impl RpcHandlerError for RpcLightClientNextBlockError {
+    fn parse_raw_error(value: serde_json::Value) -> Option<Result<Self, serde_json::Error>> {
+        common::parse_unknown_block!(value => Self)
+    }
+}
+
+impl RpcMethod for RpcLightClientNextBlockRequest {
+    type Response = RpcLightClientNextBlockResponse;
+    type Error = RpcLightClientNextBlockError;
+
+    fn method_name(&self) -> &str {
+        "next_light_client_block"
+    }
+
+    fn params(&self) -> Result<serde_json::Value, io::Error> {
+        Ok(json!(self))
+    }
+}
+
+impl private::Sealed for RpcLightClientNextBlockRequest {}
