+++ src/methods/sandbox/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+pub mod patch_state;
+pub use patch_state as sandbox_patch_state;
+
+pub mod fast_forward;
+pub use fast_forward as sandbox_fast_forward;
